# contracts/openapi_command_center.yaml
openapi: 3.1.0
info:
  title: ICS/OT Command Center API
  description: 'Read-only product API for the operator command center. Legacy contracts:
    contracts/asset_api_v1.yaml, contracts/asset_api_v2.yaml.'
  version: 3.0.0
paths:
  /health:
    get:
      summary: Health
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /diagnostics:
    get:
      summary: Diagnostics
      operationId: diagnostics_diagnostics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /data/sources:
    get:
      summary: Data Sources
      description: Canonical estate sources loaded at runtime (APP-01).
      operationId: data_sources_data_sources_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /data/views/estate:
    get:
      summary: Data Estate View
      description: "Live derived estate summary \u2014 not acceptance-test fixtures."
      operationId: data_estate_view_data_views_estate_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /data/views/estate-by-plant:
    get:
      summary: Estate By Plant View
      description: Per-plant estate aggregates for Home dashboard. Alert counts are inventory
        signals, not contextual risk rank.
      operationId: data_estate_by_plant_data_views_estate_by_plant_get
      parameters:
      - name: include_top_alerts
        in: query
        required: false
        schema:
          type: integer
          default: 3
          maximum: 10
          minimum: 0
          title: Include Top Alerts
      - name: severity_min
        in: query
        required: false
        schema:
          type: string
          title: Severity Min
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /plants:
    get:
      summary: List Plants
      description: All workshop plants (18 rows) sorted by plant_id.
      operationId: list_plants_plants_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /plants/{plant_id}/assets:
    get:
      summary: List Plant Assets
      operationId: list_plant_assets_plants__plant_id__assets_get
      parameters:
      - name: plant_id
        in: path
        required: true
        schema:
          type: string
          title: Plant Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          maximum: 500
          title: Limit
      - name: q
        in: query
        required: false
        schema:
          type: string
          title: Q
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Plant Not Found
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /assets/{asset_id}/alerts:
    get:
      summary: List Asset Alerts
      operationId: list_asset_alerts_assets__asset_id__alerts_get
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          maximum: 200
          title: Limit
      - name: severity
        in: query
        required: false
        schema:
          type: string
          title: Severity
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Asset Not Found
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /data/views/vendor-sessions:
    get:
      summary: Data Vendor Sessions
      description: Vendor/remote access sessions with anomaly flags.
      operationId: data_vendor_sessions_data_views_vendor_sessions_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /assets/{id}/identity:
    get:
      summary: Asset Identity
      operationId: asset_identity_assets__id__identity_get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /identity/conflicts:
    get:
      summary: Identity Conflicts
      operationId: identity_conflicts_identity_conflicts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /telemetry/quality:
    get:
      summary: Telemetry Quality
      operationId: telemetry_quality_telemetry_quality_get
      parameters:
      - name: plant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
      - name: asset_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /telemetry/timeline:
    get:
      summary: Telemetry Timeline
      operationId: telemetry_timeline_telemetry_timeline_get
      parameters:
      - name: tag_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag Id
      - name: plant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
      - name: asset_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
      - name: order
        in: query
        required: false
        schema:
          type: string
          default: event_time
          title: Order
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 200
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /risk/contextual:
    get:
      summary: Risk Contextual
      operationId: risk_contextual_risk_contextual_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /safety/conflicts:
    get:
      summary: Safety Conflicts
      operationId: safety_conflicts_safety_conflicts_get
      parameters:
      - name: plant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Plant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /recovery/{site_or_unit}:
    get:
      summary: Recovery View
      operationId: recovery_view_recovery__site_or_unit__get
      parameters:
      - name: site_or_unit
        in: path
        required: true
        schema:
          type: string
          title: Site Or Unit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /authority/actions:
    get:
      summary: Authority Actions
      operationId: authority_actions_authority_actions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /graph/slice:
    get:
      summary: Graph Slice View
      operationId: graph_slice_view_graph_slice_get
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
          default: Q5
          title: Query
      - name: asset_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset Id
      - name: plant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Plant Id
      - name: alert_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Alert Id
      - name: alias
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Alias
      - name: hop_cap
        in: query
        required: false
        schema:
          type: integer
          default: 8
          title: Hop Cap
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ops/slo:
    get:
      summary: Ops Slo
      operationId: ops_slo_ops_slo_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /ops/cost-per-incident:
    get:
      summary: Ops Cost Per Incident
      operationId: ops_cost_per_incident_ops_cost_per_incident_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /recommend:
    post:
      summary: Recommend
      operationId: recommend_recommend_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /eval/run:
    post:
      summary: Eval Run
      description: "Local harness only \u2014 runs golden eval cases (ENH-09). No\
        \ OT I/O."
      operationId: eval_run_eval_run_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/EvalRunRequest'
              - type: 'null'
              title: Request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /audit/traces:
    get:
      summary: Audit Traces
      description: Decision trace list for audit screen (APP-02).
      operationId: audit_traces_audit_traces_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /scenarios/catalog:
    get:
      summary: Scenario Catalog
      description: "APP-03 golden scenario rail \u2014 plant/asset/alert and expected\
        \ badges."
      operationId: scenario_catalog_scenarios_catalog_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /scenarios/{scenario_id}:
    get:
      summary: Get Scenario
      description: Scenario binding plus optional timeline JSON fixture.
      operationId: get_scenario_scenarios__scenario_id__get
      parameters:
      - name: scenario_id
        in: path
        required: true
        schema:
          type: string
          title: Scenario Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /data/shift-notes/untrusted:
    get:
      summary: Shift Notes Untrusted
      description: "Untrusted shift handover \u2014 never treated as authoritative. Optional plant_id is request scope only; source has no plant_id (OPEN-018)."
      operationId: shift_notes_untrusted_data_shift_notes_untrusted_get
      parameters:
        - name: plant_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /agent/workflow/demo:
    get:
      summary: Agent Workflow Demo
      description: Expose workflow state shape for UI / EVAL-016 (deterministic demo
        envelope).
      operationId: agent_workflow_demo_agent_workflow_demo_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    EvalRunRequest:
      properties:
        case_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Case Ids
          description: "Optional subset; default all EVAL-001\u2026031"
      type: object
      title: EvalRunRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    RecommendRequest:
      properties:
        actor:
          type: string
          title: Actor
        purpose:
          type: string
          title: Purpose
        plant_id:
          type: string
          title: Plant Id
        as_of:
          type: string
          title: As Of
          default: workshop-static
        policy_version:
          type: string
          title: Policy Version
          default: policy.py:ACTION_TIERS
        asset_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset Id
        alert_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Alert Id
        severity:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
        process_context:
          anyOf:
          - type: string
          - type: 'null'
          title: Process Context
      type: object
      required:
      - actor
      - purpose
      - plant_id
      title: RecommendRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError